end tasks

+++ src/main/java/com/example/listingapp/servicies/CategoryService.java
@@ -2,16 +2,19 @@ package com.example.listingapp.servicies;
 
 import com.example.listingapp.model.Category;
 import com.example.listingapp.repositories.CategoryRepository;
+import com.example.listingapp.repositories.ListingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Locale;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final ListingRepository listingRepository;
 
     public List<Category> findAll() {
         return categoryRepository.findAll();
@@ -22,10 +25,21 @@ public class CategoryService {
     }
 
     public Category save(Category category) {
+        Optional<Category> byName = categoryRepository.findByName(category.getName().toUpperCase(Locale.ROOT));
+        if (byName.isPresent()) {
+            return byName.get();
+        }
+        category.setName(category.getName().toUpperCase(Locale.ROOT));
         return categoryRepository.save(category);
     }
 
-    public void deleteById(int id) {
+    public boolean deleteById(int id) {
+        Optional<Category> byId = categoryRepository.findById(id);
+        if (byId.isEmpty()) {
+            return false;
+        }
+        listingRepository.changeAllListingsCategory(byId.get().getId(), null);
         categoryRepository.deleteById(id);
+        return true;
     }
 }
